Isolated Module
This patch changes the design of this project such that the relationship
between a host program and plug-in is now unidirectional -- the plug-in
never attempts to access or rely on types in the host program. The type
registration functionality is now provided via a symbol all plug-ins are
expected to have, `Types map[string]func()interface{}`. This symbol is a
type map that the host program uses to register the plug-in's modules.